Reduce complexity of delete redflag function [#162623761]  On branch reduce-complexity  Changes to be committed: 	modified:   .coverage 	modified:   api/app.py 	new file:   api/routes/delete_incident.py 	modified:   api/routes/incidents.py
