+++ api/routes/incidents.py
@@ -123,65 +123,3 @@ def get_a_red_flag(red_flag_id):
         jsonify({"status": 404, "error": "Red-flag record does not exist"}),
         404,
     )
-
-
-@red_flags_bp.route("/red-flags/<red_flag_id>", methods=["DELETE"])
-@token_required
-@non_admin
-def delete_record(red_flag_id):
-    record_id = red_flag_id
-    if not is_valid_id(record_id):
-        return (
-            jsonify({"error": "Red-flag id must be an number", "status": 400}),
-            400,
-        )
-
-    incident_id = int(red_flag_id)
-    results = get_incident_obj_by_id(
-        incident_id=incident_id, collection=red_flags
-    )
-
-    response = None
-    if not results:
-        response = (
-            jsonify(
-                {"status": 404, "error": "Red-flag record does not exist"}
-            ),
-            404,
-        )
-    elif not results.created_by == get_current_identity():
-
-        response = (jsonify({"status": 403, "error": delete_not_allowed}), 403)
-    elif results.status == "draft":
-        record_id = results.incident_id
-        record_index = red_flags.index(results)
-        red_flags.pop(record_index)
-
-        response = (
-            jsonify(
-                {
-                    "status": 200,
-                    "data": [
-                        {
-                            "id": record_id,
-                            "message": "red-flag record has been deleted",
-                        }
-                    ],
-                }
-            ),
-            200,
-        )
-    else:
-        response = (
-            jsonify(
-                {
-                    "status": 403,
-                    "error": (
-                        "You are not allowed to delete a red-flag which is"
-                        f" {results.status }"
-                    ),
-                }
-            ),
-            403,
-        )
-    return response
